// backend/template/src/app/services/models/project-user.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Project } from '../models/project';
import { ProjectUserId } from '../models/project-user-id';
import { User } from '../models/user';
export interface ProjectUser {
  created_at?: string;
  id?: ProjectUserId;
  project?: Project;
  role?: 'MANAGER' | 'TEAM_MEMBER';
  updated_at?: string;
  user?: User;
}